requirements.txt
pillow==11.1.0
streamlit==1.43.0
streamlit-cropper==0.2.2
numpy==2.0.2
opencv-python==4.11.0.86
keras==3.9.0
pandas==2.2.3
torch==2.6.0+cpu
scikit-image==0.25.2